--- a/lib/http_api_tools/model/attributes.rb
+++ b/lib/http_api_tools/model/attributes.rb
@@ -23,7 +23,10 @@ module HttpApiTools
         attrs = attrs.with_indifferent_access if self._with_indifferent_access
 
         attributes.each do |attr_name, attr_options|
-          raw_value = attrs[attr_name] == nil ? default_for(attr_options) : attrs[attr_name]
+          value = attrs[attr_name]
+          value = attrs[attr_name.to_s.freeze] if value.nil? && !self._with_indifferent_access
+
+          raw_value = value == nil ? default_for(attr_options) : value
           set_raw_value(attr_name, raw_value, true) unless raw_value == nil
         end