--- a/spec/http_api_tools/support/spec_models_for_serializing.rb
+++ b/spec/http_api_tools/support/spec_models_for_serializing.rb
@@ -46,6 +46,8 @@ class Company
   belongs_to :parent_company
   belongs_to :address
 
+  with_indifferent_access(true)
+
 
   #Act like active record for reflectively interogating type info
   def self.reflections