## [Master](https://buildkite.com/hooroo/hat-dot-gem/builds/11)
### [JRuby](https://buildkite.com/hooroo/hat-dot-gem/builds/11#6e212d98-0657-4c48-9fcd-ea68813d6ed3/95-136)
```
HttpApiTools::Nesting::JsonSerializer
nested-models: 17854.0 ms
nested-serializes: 42975.0 ms
nested-json: 881.0 ms
 
HttpApiTools::Sideloading::JsonSerializer
sideloaded-models: 16685.0 ms
sideloaded-serializes: 111.0 ms
sideloaded-json: 5.0 ms
```

### [MRI](https://buildkite.com/hooroo/hat-dot-gem/builds/11#9847ad6b-83bc-4934-911a-2451de52566f/95-100)
```
HttpApiTools::Nesting::JsonSerializer
nested-models: 31110.592909000003 ms
nested-serializes: 66908.016696 ms
nested-json: 2273.230271 ms
 
HttpApiTools::Sideloading::JsonSerializer
sideloaded-models: 33484.473164 ms
sideloaded-serializes: 167.360278 ms
sideloaded-json: 5.7206790000000005 ms
```

## [bili_perf](https://buildkite.com/hooroo/hat-dot-gem/builds?branch=bili_perf)
### [JRuby](https://buildkite.com/hooroo/hat-dot-gem/builds/17#c8f55aab-aca5-445c-83c3-0b1f23963311/97-138)
```
HttpApiTools::Nesting::JsonSerializer
nested-models: 4280.0 ms
nested-serializes: 8417.0 ms
nested-json: 727.0 ms
 
HttpApiTools::Sideloading::JsonSerializer
sideloaded-models: 3519.0 ms
sideloaded-serializes: 104.0 ms
sideloaded-json: 5.0 ms
```

### [MRI](https://buildkite.com/hooroo/hat-dot-gem/builds/17#59fe45ad-9a6a-4596-a86b-e1c55cc99f5b/97-102)
```
HttpApiTools::Nesting::JsonSerializer
nested-models: 9625.82165 ms
nested-serializes: 24572.071655 ms
nested-json: 2480.702499 ms
 
HttpApiTools::Sideloading::JsonSerializer
sideloaded-models: 10332.308903 ms
sideloaded-serializes: 110.826937 ms
sideloaded-json: 5.725747 ms
```


Some of the biggest wins were:
* Defining the root key at a class level (plurualize and singularize at an instance level was performing many regex)
* Remove `with_indifferent_access`
* BigDecimal == nil is much slower than BigDecimal.nil?


## Nested Performance Increase
JRuby Model Creation: 4x
JRuby Model Serialization: 5x

MRI Model Creation: 3x
MRI Model Serialization: 2x