@robmonie 2016-01-13
I haven't looked at the code in detail but the results of this look great!